+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Yii 2 Content Management extension Change Log
 -----------------------
 
 - Bug #1: Fixed `Manager::getMetaData()` produces E_WARNING for the item, which is not present at source storage (klimov-paul)
+- Enh #2: Created `yii2tech\content\mail\MailerContentBehavior` for mail composition simplification (klimov-paul)
 - Enh: Usage of deprecated `yii\base\InvalidParamException` replaced by `yii\base\InvalidArgumentException` (klimov-paul)